fix warningow i up dokumentacji

+++ Snake SFML 2.x/src/Biblioteki.cpp
@@ -6,7 +6,7 @@
 int skinWeza = 0, Wynik = 0;
 float glosnoscMuzyki = 15.0f;
 
-void losuj(float& zmienna, int parametr) { zmienna = rand() % parametr; }
+void losuj(float& zmienna, int parametr) { zmienna =(float)( rand() % parametr); }
 int losuj(int parametr) { return rand() % parametr; }
 
 bool kolizja(Sprite s1, Sprite s2) { return s1.getGlobalBounds().intersects(s2.getGlobalBounds()); }
